Add checkbox to toggle combining veteran groups
Limit the columns to max-content and add a gap and padding, which are removed for printing.
Don't show the loading message if we already have results data and it's just refetching due to the cache expiring.
Add hasVeteranPref to all applicants, so they get the asterisk in all columns.
Make toolbar sticky.
Tweak header h3 sizes.